+++ packages/logger/src/tracing/transports/http.ts
@@ -1,6 +1,8 @@
 import type { TraceData, TraceTransport } from '@notjustcoders/one-logger-types';
 import { sdk } from '../../sdk.js';
+import { error } from 'console';
 
+import { serializeError } from 'serialize-error';
 /**
  * HTTP transport for sending traces to a backend server using the One Logger SDK
  */
@@ -34,6 +36,7 @@ export class HttpTraceTransport implements TraceTransport {
           startTime: span.startTime,
           endTime: span.endTime,
           duration: span.duration,
+          error: span.error ? serializeError(span.error):span.error,
           status: span.status,
           metadata: span.metadata || {}
         })) || []